+++ index.html
@@ -457,6 +457,6 @@
         <button type="button" class="error__button">Попробовать снова</button>
       </div>
     </template>
-    <script type="module" src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
   </body>
 </html>
