+++ js/const.js
@@ -1,0 +1,49 @@
+const QUANTITY = 10;
+const MIN_GUESTS = 1;
+const MAX_GUESTS = 3;
+const MIN_PRICE = 0;
+const MAX_PRICE = 1000000;
+const LINK = 'img/avatars/user0';
+const Y_MIN = 139.7;
+const Y_MAX = 139.8;
+const X_MIN = 35.65;
+const X_MAX = 35.7;
+const TYPES_OF_RENT = ['palace', 'flat', 'house', 'bungalo'];
+const ROOMS_QUANTITY = [1, 2, 3, 100];
+const CHECK_HOURS = ['12:00', '13:00', '14:00'];
+
+const FEATURES_LIST = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTOS_LIST = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+const offersList = [];
+//prettier-ignore
+export {
+  QUANTITY,
+  MIN_GUESTS,
+  MAX_GUESTS,
+  MIN_PRICE,
+  MAX_PRICE,
+  LINK,
+  Y_MIN,
+  Y_MAX,
+  X_MIN,
+  X_MAX,
+  TYPES_OF_RENT,
+  ROOMS_QUANTITY,
+  CHECK_HOURS,
+  FEATURES_LIST,
+  PHOTOS_LIST,
+  offersList
+}
