--- js/random-array-item.js
@@ -1,5 +1,0 @@
-const getRandomArrayItem = function (arr) {
-  return arr[Math.floor(Math.random() * arr.length)];
-};
-
-export { getRandomArrayItem };
