--- js/random-number.js
@@ -1,6 +1,0 @@
-const getRandomNumber = function (min, max, decimalPoints) {
-  const findRandom = Math.random() * (max - min) + min;
-  return +findRandom.toFixed(decimalPoints);
-};
-
-export { getRandomNumber };
